% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.tada.list}
\alias{as.tada.list}
\title{Utility functions for textmatch objects}
\usage{
as.tada.list(x)
}
\arguments{
\item{x}{a textmatch list of distances}
}
\value{
a data frame
}
\description{
Utility functions for textmatch objects
}
