% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pair_distances.R
\name{calculate_pair_distances}
\alias{calculate_pair_distances}
\title{Create a data frame of pairs of documents obtained through
coarsened exact matching (CEM) within a specified number of bins
and return indices for matched sets}
\usage{
calculate_pair_distances(rep.list, Z, propensity.method = NULL,
  include = c("cosine", "euclidean", "mahalanobis", "propensity"),
  exclude = c("jaccard"))
}
\arguments{
\item{rep.list}{a named list of representations}

\item{Z}{A logical or binary vector indicating treatment and control for each unit in the study.
TRUE or 1 represents a treatment unit, FALSE of 0 represents a control unit.}

\item{propensity.method}{Either GLM or MNIR for propensity score estimation}
}
\value{
A \link{data.frame} of indices for matched pairs of documents
}
\description{
Create a data frame of pairs of documents obtained through
coarsened exact matching (CEM) within a specified number of bins
and return indices for matched sets
}
