% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corp_contrast.R
\name{corp_contrast}
\alias{corp_contrast}
\title{Corpus contrasts}
\usage{
corp_contrast(corp, Z)
}
\arguments{
\item{corp}{a fitted \code{\link[stm]{stm}} object}

\item{Z}{an indicator for treatment assignment}
}
\value{
Corpus contrast summaries
}
\description{
Corpus contrasts
}
