% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{FoxCNNcorpus}
\alias{FoxCNNcorpus}
\title{Corpus with 1,565 articles from CNN and 1,796 articles from Fox News.}
\format{A \link{data.frame} with 3,361 observations of 5 articles}
\usage{
FoxCNNcorpus
}
\description{
Corpus with 1,565 articles from CNN and 1,796 articles from Fox News.
}
\details{
Corpus of front-page news articles published online by CNN or Fox News
from 12/20/2014 to 05/09/2015 containing 1,565 articles from CNN and 1,796 articles from Fox News.
Data contains article identifiers corresponding to data in FoxCNNmeta as well as raw and cleaned text data.
}
\references{
Mozer et al. (2019)
 \href{https://reaganmozer.com/textmatching}{Matching with Text Data: An Experimental Evaluation of Methods for Matching Documents
 and of Measuring Match Quality}. \emph{Political Analysis}, Forthcoming.
}
\keyword{data}
