% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{FoxCNNmeta}
\alias{FoxCNNmeta}
\title{Dataset containing metadata for FoxCNN corpus with 1,565 articles from CNN and 1,796 articles from Fox News.}
\format{
A \link{data.frame} with 3,361 observations of 5 articles
}
\usage{
FoxCNNmeta
}
\description{
Dataset containing metadata for FoxCNN corpus with 1,565 articles from CNN and 1,796 articles from Fox News.
}
\details{
Metadata for the FoxCNN corpus including article names, original URLs, and dates of publication.
}
\references{
Mozer et al. (2019)
 \href{https://reaganmozer.com/textmatching}{Matching with Text Data: An Experimental Evaluation of Methods for Matching Documents
 and of Measuring Match Quality}. \emph{Political Analysis}, Forthcoming.
}
\keyword{data}
