% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{FoxCNNscores}
\alias{FoxCNNscores}
\title{Dataset containing quality scores for 2.8 million
possible matched pairs of Fox and CNN articles corresponding to 117 different combinations
of distance metric and text representation, enumerated in Mozer et al. (2019), Section 3.
Pair quality scores are estimated from a predictive model trained on a sample of 505 pairs
evaluated by human coders.}
\format{A \link{data.frame} with 2,810,740 pairs and 117 features.}
\usage{
FoxCNNscores
}
\description{
Dataset containing quality scores for 2.8 million
possible matched pairs of Fox and CNN articles corresponding to 117 different combinations
of distance metric and text representation, enumerated in Mozer et al. (2019), Section 3.
Pair quality scores are estimated from a predictive model trained on a sample of 505 pairs
evaluated by human coders.
}
\details{
Aggregated quality scores for all 2.8 million possible pairs of Fox and CNN articles
based on the predictive model trained on ratings from human coders.
}
\references{
Mozer et al. (2019)
 \href{https://reaganmozer.com/textmatching}{Matching with Text Data: An Experimental Evaluation of Methods for Matching Documents
 and of Measuring Match Quality}. \emph{Political Analysis}, Forthcoming.
}
\keyword{data}
