% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{FoxCNNsurvey}
\alias{FoxCNNsurvey}
\title{Dataset containing distance measurements and average quality scores
for sample of 505 pairs of matched Fox and CNN articles evaluated by human
coders on Mechanical Turk. Raw quality scores are regressed on distance measurements
to fit a predictive model for match quality as a function of the 117 metrics considered.}
\format{A \link{data.frame} with 505 pairs and 108 features.}
\usage{
FoxCNNsurvey
}
\description{
Dataset containing distance measurements and average quality scores
for sample of 505 pairs of matched Fox and CNN articles evaluated by human
coders on Mechanical Turk. Raw quality scores are regressed on distance measurements
to fit a predictive model for match quality as a function of the 117 metrics considered.
}
\details{
Distance measurements and average quality scores for a sample of 505
matched pairs of documents evaluated by human coders.
}
\references{
Mozer et al. (2019)
 \href{https://reaganmozer.com/textmatching}{Matching with Text Data: An Experimental Evaluation of Methods for Matching Documents
 and of Measuring Match Quality}. \emph{Political Analysis}, Forthcoming.
}
\keyword{data}
