% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{FoxCNNtopics}
\alias{FoxCNNtopics}
\title{Dataset containing the final STM models chosen for the FoxCNNCorpus data with 
K=10, 30, 50, and 100 topics each.}
\format{Four \code{stmobj} model objects from \code{\link{stm}}.}
\usage{
FoxCNNtopics
}
\description{
Dataset containing the final STM models chosen for the FoxCNNCorpus data with 
K=10, 30, 50, and 100 topics each.
}
\details{
Output containing the final STM models chosen for the FoxCNNCorpus data with 
K=10, 30, 50, and 100 topics each.
}
\references{
Mozer et al. (2019)
\\"Matching with Text Data: An Experimental Evaluation of Methods for Matching Documents
 and of Measuring Match Quality\\". \emph{Political Analysis}, Forthcoming.
}
\keyword{data}
