% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bal.R
\name{get_bal_optmatch}
\alias{get_bal_optmatch}
\title{Generates balance diagnostics for many possible matched datasets
which can be used to make comparisons between matching methods.}
\usage{
get_bal_optmatch(matchobj, Z, covs)
}
\arguments{
\item{matchobj}{an \link{optmatch} object or a matched dataset}

\item{Z}{a vector of treatment indicators}

\item{covs}{a matrix or data frame of covariates to assess balance on}
}
\value{
A \link{data.frame} MNIR sufficient reduction scores for a corpus
}
\description{
Generates balance diagnostics for many possible matched datasets
which can be used to make comparisons between matching methods.
}
