% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CEM.R
\name{get_CEM}
\alias{get_CEM}
\title{Create a data frame of pairs of documents obtained through
coarsened exact matching (CEM) within a specified number of bins
and return indices for matched sets}
\usage{
get_CEM(x, Z, rep.name, cuts = NULL, caliper_fun = NULL,
  verbose = FALSE, SR = NULL)
}
\arguments{
\item{x}{a text representation}

\item{Z}{a vector of treatment indicators}

\item{rep.name}{a string or character with the name of the matching method}

\item{cuts}{a function for how the variables will be binned. Defaults to "median"}
}
\value{
A \link{data.frame} of indices for matched pairs of documents
}
\description{
Create a data frame of pairs of documents obtained through
coarsened exact matching (CEM) within a specified number of bins
and return indices for matched sets
}
