% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matches.R
\name{get_matches}
\alias{get_matches}
\title{Similarity and distance computation between documents or features}
\usage{
get_matches(x, Z, dist.name, caliper_fun, text_caliper = 0.001, tol = 0)
}
\arguments{
\item{x}{a matrix of pairwise distances for all potential matches of treatment and control units. See \link{pair_distances}.}

\item{Z}{a vector of treatment indicators}

\item{dist.name}{a string or character with the name of the matching method}

\item{caliper_fun}{an optional function specifying the caliper to enforce when matching}

\item{text_caliper}{should a text-based caliper be enforced? If so, specify the caliper (as a quantile of the distribution of pairwise document distances)}
}
\description{
These functions compute distance matrices from a text representation where each row is a document
and each column is a feature to measure distance over based on treatment indicator Z
}
