% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_scores.R
\name{get_similarity_scores}
\alias{get_similarity_scores}
\title{This function calculates an input character vector's similarity matrix according to the measures contained in the predictive model.}
\usage{
get_similarity_scores(x)
}
\arguments{
\item{x}{A character vector where each element is a document}
}
\value{
A data frame of rows (n * n-1) and columns 16; each column is one of the constituent similarity measures
}
\description{
This function calculates an input character vector's similarity matrix according to the measures contained in the predictive model.
}
