% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_word2vec.R
\name{get_word2vec}
\alias{get_word2vec}
\title{This function calculates the Word2Vec embeddings}
\usage{
get_word2vec(dat)
}
\arguments{
\item{dat}{FoxCNN corpus to calculate Word2Vec scores for}
}
\value{
A vector of predicted match quality scores on a scale of 0-10.
}
\description{
This function calculates the Word2Vec embeddings
}
