% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_word2vec_glove.R
\name{get_word2vec_glove}
\alias{get_word2vec_glove}
\title{This function calculates the Word2Vec embeddings}
\usage{
get_word2vec_glove(dir.source, corpus)
}
\arguments{
\item{dat}{FoxCNN corpus to calculate Word2Vec scores for}
}
\value{
A list of data frames containing the Word2Vec projections of the corpus
}
\description{
This function calculates the Word2Vec embeddings
}
