% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matches.R
\name{makeMatches}
\alias{makeMatches}
\title{Create a data frame of matched pairs of documents and return indices for matched sets}
\usage{
makeMatches(match.obj, Z)
}
\arguments{
\item{match.obj}{a matched data set}

\item{Z}{a vector of treatment indicators}
}
\value{
A \link{data.frame} of indices for matched pairs of documents
}
\description{
Create a data frame of matched pairs of documents and return indices for matched sets
}
