% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textPlot.R
\name{norm_stats}
\alias{norm_stats}
\title{Calculate differences in means and calculate simultaneous confidence intervals
with automatic adjustment for multiple comparisons.}
\usage{
norm_stats(x, Z, alpha)
}
\arguments{
\item{x}{matrix of LIWC features}

\item{Z}{vector of treatment indicators.}
}
\value{
a \pkg{data.frame} with results.
}
\description{
Calculate differences in means and calculate simultaneous confidence intervals
with automatic adjustment for multiple comparisons.
}
