% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_distances.R
\name{pair_distances}
\alias{pair_distances}
\title{Similarity and distance computation between documents or features}
\usage{
pair_distances(dat, Z, include = c("cosine", "euclidean", "mahalanobis"),
  form = "data.frame", verbose = FALSE)
}
\arguments{
\item{include}{Which distances to calculate}

\item{form}{Should the distances be returned as a list of matrices or condensed into a single data frame?}

\item{x}{a matrix text representation with rows corresponding to each document in a corpus and columns
that represent summary measures of the text (e.g., word counts, topic proportions, etc.). Acceptable forms include
a valid \pkg{quanteda} \code{dfm} object, a \pkg{tm} Document-Term Matrix, a matrix of estimated topic proportions, 
or a vector of estimated propensity scores.}
}
\value{
A matrix showing pairwise distances for all potential matches of treatment and control units under various distance metrics
}
\description{
These functions compute distance matrices from a text representation where each row is a document
and each column is a feature to measure distance over based on treatment indicator Z
}
