% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_word2vec.R
\name{proc_pretrained_vec}
\alias{proc_pretrained_vec}
\title{This function calculates the Word2Vec embeddings}
\usage{
proc_pretrained_vec(p_vec)
}
\arguments{
\item{p_vec}{A character vector or subset of the FoxCNN corpus where each element is a document}
}
\value{
A vector of predicted match quality scores on a scale of 0-10.
}
\description{
This function calculates the Word2Vec embeddings
}
\keyword{internal}
