% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{quality_model}
\alias{quality_model}
\title{Fitted model for pairwise match quality as a function of 117 distance metrics
calculated in Mozer et al. (2019). Trained on "FoxCNNsurvey" dataset.}
\format{A \code{\link{glmnet}} model object.}
\usage{
quality_model
}
\description{
Fitted model for pairwise match quality as a function of 117 distance metrics
calculated in Mozer et al. (2019). Trained on "FoxCNNsurvey" dataset.
}
\details{
Fitted model for predicting the match quality score for a given
pair of text documents as a function of 117 distance measurements.
}
\references{
Mozer et al. (2019)
\\"Matching with Text Data: An Experimental Evaluation of Methods for Matching Documents
 and of Measuring Match Quality\\". \emph{Political Analysis}, Forthcoming.
}
\keyword{datasets}
