% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textPlot.R
\name{rs_stats}
\alias{rs_stats}
\title{Calculate differences in medians and perform simultaneous rank-sum tests.}
\usage{
rs_stats(x, Z)
}
\arguments{
\item{x}{matrix of LIWC features}

\item{Z}{vector of treatment indicators.}
}
\value{
a \pkg{data.frame} with results.
}
\description{
Calculate differences in medians and perform simultaneous rank-sum tests.
}
