% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SE}
\alias{SE}
\title{Utility functions for textmatch objects}
\usage{
SE(v, z)
}
\arguments{
\item{v}{a vector of covariate or outcome data}
}
\value{
z a vector of treatment indicators
}
\description{
Utility functions for textmatch objects
}
