% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_matches.R
\name{select_matches}
\alias{select_matches}
\title{Given a model for match quality and a corpus of documents, calculate the 
estimated match quality of each potential pairing of treated and control documents
and return a matched dataset containing all pairs of documents with estimated quality
above a specified threshold}
\usage{
select_matches(corpus, Z, mod, threshold)
}
\arguments{
\item{corpus}{description}

\item{Z}{treatment indicator}

\item{threshold}{for quality scores to return}
}
\value{
A \link{data.frame} of matched pairs of documents
}
\description{
Given a model for match quality and a corpus of documents, calculate the 
estimated match quality of each potential pairing of treated and control documents
and return a matched dataset containing all pairs of documents with estimated quality
above a specified threshold
}
