% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tada.R
\name{tada}
\alias{tada}
\title{Extract an array of numerical features from a collection of text documents}
\usage{
tada(text, lower = TRUE, alphaNum = TRUE, stopwords = NULL, stem = FALSE)
}
\arguments{
\item{text}{a character vector of text documents}

\item{control}{a named list of control options for evaluating documents individually and globally}
}
\value{
a \code{data.frame} representation of text with columns corresponding to text-based features
}
\description{
Extract an array of numerical features from a collection of text documents
}
