% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmatch.R
\name{textmatch}
\alias{textmatch}
\title{This function runs the main ML model as specified in Mozer et al. (2018)}
\usage{
textmatch(x, outcome = "matrix", verbose = TRUE)
}
\arguments{
\item{x}{A character vector or subset of the FoxCNN corpus where each element is a document}
}
\value{
A vector of predicted match quality scores on a scale of 0-10.
}
\description{
This function runs the main ML model as specified in Mozer et al. (2018)
}
