% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textPlot.R
\name{textPlot}
\alias{textPlot}
\title{Plot differences between groups of documents on psychological dimensions measured by LIWC.}
\usage{
textPlot(
  x,
  Z,
  sig.only = FALSE,
  p.adj = NULL,
  alpha = 0.05,
  main = NULL,
  test.stat = NULL
)
}
\arguments{
\item{x}{matrix of LIWC features}

\item{Z}{vector of treatment indicators.}
}
\value{
a \pkg{plotrix} plot.
}
\description{
Plot differences between groups of documents on psychological dimensions measured by LIWC.
}
