% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textPS.R
\name{textPS_dist}
\alias{textPS_dist}
\title{Given a text representation x and treatment indicator Z,
construct a distance matrix for the pairwise distances between
treatment and control documents based on the textual propensity score.}
\usage{
textPS_dist(x, Z, normalize = TRUE, return.df = TRUE,
  verbose = FALSE)
}
\arguments{
\item{x}{a TDM or DFM text representation}

\item{Z}{a vector of treatment indicators}

\item{normalize}{should the TDM features be normalized? Defaults to TRUE.}

\item{return.df}{Should the distances be returned as a data frame? Default is TRUE.}
}
\description{
Given a text representation x and treatment indicator Z,
construct a distance matrix for the pairwise distances between
treatment and control documents based on the textual propensity score.
}
