% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_dfm.R
\name{transform_dfm}
\alias{transform_dfm}
\title{Applies bounds, weights, and/or coarsening schemes to a dfm or document frequency matrix
to reduce the dimension of the data, reduce noise, or apply other design rules
(e.g. - to exclude words that occur in too few or too many documents).}
\usage{
transform_dfm(x, bounds, tfidf = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{a matrix text representation with rows corresponding to each document in a corpus and columns
that represent summary measures of the text (e.g., word counts, topic proportions, etc.). Acceptable forms include
a valid \pkg{quanteda} \code{dfm} object, a \pkg{tm} Document-Term Matrix, or a matrix of estimated topic proportions.}

\item{bounds}{a vector of lower and upper bounds to enforce. Defaults to excluding any terms that appear in only one document
and any terms that appear in every document}

\item{tfidf}{optional scheme to use for weighting the DTM. Defaults to \code{FALSE}.}

\item{verbose}{indicator for verbosity}
}
\value{
A bounded DFM
}
\description{
Applies bounds, weights, and/or coarsening schemes to a dfm or document frequency matrix
to reduce the dimension of the data, reduce noise, or apply other design rules
(e.g. - to exclude words that occur in too few or too many documents).
}
