% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MNIR.R
\name{get_MNIR}
\alias{get_MNIR}
\title{Create a data frame of pairs of documents obtained through
coarsened exact matching (CEM) within a specified number of bins
and return indices for matched sets}
\usage{
get_MNIR(x, Z, weighting = NULL, rep.name = NULL)
}
\arguments{
\item{x}{a TDM or DFM text representation}

\item{Z}{a vector of treatment indicators}
}
\value{
A \link{data.frame} MNIR sufficient reduction scores for a corpus
}
\description{
Create a data frame of pairs of documents obtained through
coarsened exact matching (CEM) within a specified number of bins
and return indices for matched sets
}
