% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmatch.R
\name{textmatch}
\alias{textmatch}
\title{This function runs the main ML model as specified in Mozer et al. (2018)}
\usage{
textmatch(x, outcome = "matrix")
}
\arguments{
\item{x}{A character vector where each element is a document}
}
\value{
An n by n matrix where n is the length of parameter x. Each entry is a standardized similarity score.
}
\description{
This function runs the main ML model as specified in Mozer et al. (2018)
}
\examples{
textmatch(c("I am a dog", "I am a cat", "The rain in Spain falls mainly on the plain."),
 outcome = "matrix")

}
