% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MNIR.R, R/textPS.R
\name{textPS}
\alias{textPS}
\title{Calculate a univariate projection and pairwise distances for a corpus of text documents 
with binary treatment indicator Z.}
\usage{
textPS(x, Z, verbose = FALSE, g = 1)

textPS(x, Z, verbose = FALSE, g = 1)
}
\arguments{
\item{x}{a TDM or DFM text representation}

\item{Z}{a vector of treatment indicators}

\item{normalize}{should the TDM features be normalized? Defaults to TRUE.}

\item{return.df}{Should the distances be returned as a data frame? Default is TRUE.}

\item{x}{a TDM or DFM text representation}

\item{Z}{a vector of treatment indicators}
}
\value{
A \link{data.frame} MNIR sufficient reduction scores for a corpus

A \link{data.frame} MNIR sufficient reduction scores for a corpus
}
\description{
Calculate a univariate projection and pairwise distances for a corpus of text documents 
with binary treatment indicator Z.

Given a text representation x and treatment indicator Z,
estimate a univariate projection of the text data that can serve
as a sufficient statistic for the confounding information captured within the text.
}
